--- a/internal.h
+++ b/internal.h
@@ -2124,6 +2124,7 @@ char *rb_str_to_cstr(VALUE str);
 VALUE rb_str_eql(VALUE str1, VALUE str2);
 VALUE rb_obj_as_string_result(VALUE str, VALUE obj);
 const char *ruby_escaped_char(int c);
+VALUE rb_str_opt_plus(VALUE, VALUE);
 
 /* expect tail call optimization */
 static inline VALUE