@k0kubun 2019-08-06
Looks great. Thank you!

```yml
$ cat bench.yml
prelude: |
  def three
    1 + 2
  end
benchmark:
  - three
```
```
$ benchmark-driver --rbenv 'before;after;before --jit;after --jit' -v --run-duration=10 --runner=block bench.yml
before: ruby 2.7.0dev (2019-08-06T11:15:07Z master 1ee88c51b3) [x86_64-linux]
after: ruby 2.7.0dev (2019-08-06T04:25:14Z gh-2321 301f22ef4b) [x86_64-linux]
before --jit: ruby 2.7.0dev (2019-08-06T11:15:07Z master 1ee88c51b3) +JIT [x86_64-linux]
after --jit: ruby 2.7.0dev (2019-08-06T04:25:14Z gh-2321 301f22ef4b) +JIT [x86_64-linux]
Warming up --------------------------------------
               three    21.158M i/s -     70.636M times in 3.338532s (47.26ns/i, 203clocks/i)
Calculating -------------------------------------
                         before       after  before --jit  after --jit
               three    28.943M     26.677M       37.719M      41.797M i/s -    211.577M times in 7.310003s 7.931167s 5.609218s 5.061972s

Comparison:
                            three
         after --jit:  41797308.3 i/s
        before --jit:  37719481.4 i/s - 1.11x  slower
              before:  28943462.1 i/s - 1.44x  slower
               after:  26676628.6 i/s - 1.57x  slower
```