@ko1 2019-08-07
```
Comparison:
                            three
         after --jit:  41797308.3 i/s
        before --jit:  37719481.4 i/s - 1.11x  slower
              before:  28943462.1 i/s - 1.44x  slower
               after:  26676628.6 i/s - 1.57x  slower
```

why *after* slows down?