@ko1 2019-08-07
Kokubun-san edited last measurement with increasing repeat count, and it shows no slow down.